Rename module name.
