+++ cubing.py
